+++ JavaEE/src/tp5/DemoFleuriste.java
@@ -6,11 +6,11 @@ import java.util.List;
 public class DemoFleuriste {
 
 	public static void main(String[] args) {
-		Fleur fleur1 = new Fleur("Rose", (float) 6.0, 3);
-		Fleur fleur2 = new Fleur("Arum", (float) 2.1, 12);
-		Fleur fleur3 = new Fleur("Hortensia", (float) 3.5, 45);
-		Fleur fleur4 = new Fleur("Jasmin", (float) 6.2, 6);
-		Fleur fleur5 = new Fleur("Anémone", (float) 3.0, 20);
+		Fleur fleur1 = new Fleur("Rose", 6.0, 3);
+		Fleur fleur2 = new Fleur("Arum", 2.1, 12);
+		Fleur fleur3 = new Fleur("Hortensia",  3.5, 45);
+		Fleur fleur4 = new Fleur("Jasmin", 6.2, 6);
+		Fleur fleur5 = new Fleur("Anémone", 3.0, 20);
 		
 		List<Fleur> fleurs = new ArrayList<>();
 		fleurs.add(fleur1);
@@ -30,7 +30,7 @@ public class DemoFleuriste {
 		fleuriste.ajoutFleur("Anémone", 13);
 		
 		fleuriste.facturation();
-		
+		//Pas tester toutes les possibilités donc peut-etre des erreurs possibles
 	}
 	
 }
